+++ tests/plugins/widget/dnd.js
@@ -12,13 +12,37 @@
 					evt.editor.dataProcessor.writer.sortAttributes = 1;
 
 					evt.editor.widgets.add( 'testwidget', {} );
+
+					evt.editor.widgets.add( 'testwidget2', {
+						editables: {
+							n1: {
+								selector: '.n1',
+								allowedContent: 'p;blockquote(testwidget3)'
+							},
+							n2: {
+								selector: '.n2',
+								allowedContent: 'p;blockquote'
+							}
+						}
+					} );
+
+					evt.editor.widgets.add( 'testwidget3', {
+						requiredContent: 'blockquote(testwidget3)'
+					} );
+
+					evt.editor.widgets.add( 'testwidget4', {
+						editables: {
+							n1: '.n1'
+						}
+					} );
 				}
 			}
 		}
 	};
 
 	var fixHtml = widgetTestsTools.fixHtml,
-		getWidgetById = widgetTestsTools.getWidgetById;
+		getWidgetById = widgetTestsTools.getWidgetById,
+		assertRelations = lineutilsTestsTools.assertRelations;
 
 	function dropEvent( data, range ) {
 		var evt = new CKEDITOR.dom.event( {
@@ -368,6 +392,74 @@
 				} finally {
 					revert();
 				}
+			} );
+		},
+
+		'test drag and drop - block widget into nested editable (ACF)': function() {
+			var editor = this.editor,
+				html = '<div data-widget="testwidget2">' +
+					'<div class="n1">' +
+						'<p>x</p>' +
+					'</div>' +
+					'<div class="n2">' +
+						'<p>y</p>' +
+					'</div>' +
+				'</div>' +
+				'<blockquote data-widget="testwidget3" class="testwidget3" id="w3">testwidget3</blockquote>';
+
+			this.editorBot.setData( html, function() {
+				var widget = getWidgetById( editor, 'w3' ),
+					repo = editor.widgets,
+					finder = repo.finder;
+
+				// Detach dragged widget from DOM to make assertion simpler.
+				widget.wrapper.remove();
+				repo._.draggedWidget = widget;
+				finder.greedySearch();
+
+				assertRelations( editor, finder, '|<div data-widget="testwidget2"><div class="n1">|<p>x</p>|</div><div class="n2"><p>y</p></div></div>|' );
+			} );
+		},
+
+		'test drag and drop - block widget into nested editable (ACF) - no filter': function() {
+			var editor = this.editor,
+				html = '<div data-widget="testwidget4">' +
+					'<div class="n1">' +
+						'<p>x</p>' +
+					'</div>' +
+				'</div>' +
+				'<blockquote data-widget="testwidget3" class="testwidget3" id="w3">testwidget3</blockquote>';
+
+			this.editorBot.setData( html, function() {
+				var widget = getWidgetById( editor, 'w3' ),
+					repo = editor.widgets,
+					finder = repo.finder;
+
+				// Detach dragged widget from DOM to make assertion simpler.
+				widget.wrapper.remove();
+				repo._.draggedWidget = widget;
+				finder.greedySearch();
+
+				assertRelations( editor, finder, '|<div data-widget="testwidget4"><div class="n1">|<p>x</p>|</div></div>|' );
+			} );
+		},
+
+		'test drag and drop - block widget into nested editable (ACF) - self-drop': function() {
+			var editor = this.editor,
+				html = '<div data-widget="testwidget4" id="w4">' +
+					'<div class="n1">' +
+						'<p>x</p>' +
+					'</div>' +
+				'</div>';
+
+			this.editorBot.setData( html, function() {
+				var repo = editor.widgets,
+					finder = repo.finder;
+
+				repo._.draggedWidget = getWidgetById( editor, 'w4' );
+				finder.greedySearch();
+
+				assertRelations( editor, finder, '|<div data-widget="testwidget4" id="w4"><div class="n1"><p>x</p></div></div>|' );
 			} );
 		}
 	} );